Converted Floats to Doubles

+++ main.cu
@@ -8,7 +8,7 @@
 
 #include "neuralNetwork.h"
 
-float init_weight(){ return ((float) rand())/ ((float)RAND_MAX);}
+double init_weight(){ return ((double) rand())/ ((double)RAND_MAX);}
 
 
 void shuffle(int* array, size_t n){
@@ -37,46 +37,46 @@ int main(int argc, char** argv){
  	srand((unsigned)time(&t));
 	// Set the learning rate & epochs
 	int epochs = 10000;
-	float lr = 1.0f;
+	double lr = 1.0f;
 
 	int numInputs = 2;
 	int numHiddenNodes = 4;
 	int numOutputs = 1;
 
-	float training_inputs[8] = {0.0f,0.0f, 0.0f, 1.0f, 1.0f, 0.0f, 1.0f, 1.0f};
-	float training_outputs[4] = {0.0f, 1.0f, 1.0f, 0.0f};
+	double training_inputs[8] = {0.0f,0.0f, 0.0f, 1.0f, 1.0f, 0.0f, 1.0f, 1.0f};
+	double training_outputs[4] = {0.0f, 1.0f, 1.0f, 0.0f};
 	int trainingSetOrder[] = {0,1,2,3};
 
 	// Initialize all the arrays into memory
-	float* hiddenLayer = (float*) malloc(numHiddenNodes * sizeof(float));
-	float* outputLayer = (float*) malloc(numOutputs * sizeof(float));
+	double* hiddenLayer = (double*) malloc(numHiddenNodes * sizeof(double));
+	double* outputLayer = (double*) malloc(numOutputs * sizeof(double));
 
-	float* hiddenLayerBias = (float*) malloc(numHiddenNodes * sizeof(float));
-	float* outputLayerBias = (float*) malloc(numOutputs * sizeof(float));
+	double* hiddenLayerBias = (double*) malloc(numHiddenNodes * sizeof(double));
+	double* outputLayerBias = (double*) malloc(numOutputs * sizeof(double));
 
-	float* hiddenWeights = (float*)malloc(numInputs * numHiddenNodes* sizeof(float));
-	float* outputWeights = (float*)malloc(numHiddenNodes * numOutputs * sizeof(float));
+	double* hiddenWeights = (double*)malloc(numInputs * numHiddenNodes* sizeof(double));
+	double* outputWeights = (double*)malloc(numHiddenNodes * numOutputs * sizeof(double));
 
 	//cuda
-	float* cuHiddenLayer;
-	float* cuOutputLayer;
-	float* cuHiddenLayerBias;
-	float* cuOutputLayerBias;
-	float* cuOutputWeights;
-	float* cuHiddenWeights;
-	float* cuTrainingInputs;
-	float* cuTrainingOutputs;
+	double* cuHiddenLayer;
+	double* cuOutputLayer;
+	double* cuHiddenLayerBias;
+	double* cuOutputLayerBias;
+	double* cuOutputWeights;
+	double* cuHiddenWeights;
+	double* cuTrainingInputs;
+	double* cuTrainingOutputs;
 	int* cuTrainingSetOrder;
 
-	cudaMalloc((void**)&cuHiddenLayer, numHiddenNodes * sizeof(float));
-	cudaMalloc((void**)&cuOutputLayer, numOutputs * sizeof(float));
-	cudaMalloc((void**)&cuHiddenLayerBias, numHiddenNodes * sizeof(float));
-	cudaMalloc((void**)&cuOutputLayerBias, numOutputs * sizeof(float));
-	cudaMalloc((void**)&cuHiddenWeights, numInputs * numHiddenNodes * sizeof(float));
-	cudaMalloc((void**)&cuTrainingInputs, 8 * sizeof(float));
-	cudaMalloc((void**)&cuTrainingOutputs, 4 * sizeof(float));
+	cudaMalloc((void**)&cuHiddenLayer, numHiddenNodes * sizeof(double));
+	cudaMalloc((void**)&cuOutputLayer, numOutputs * sizeof(double));
+	cudaMalloc((void**)&cuHiddenLayerBias, numHiddenNodes * sizeof(double));
+	cudaMalloc((void**)&cuOutputLayerBias, numOutputs * sizeof(double));
+	cudaMalloc((void**)&cuHiddenWeights, numInputs * numHiddenNodes * sizeof(double));
+	cudaMalloc((void**)&cuTrainingInputs, 8 * sizeof(double));
+	cudaMalloc((void**)&cuTrainingOutputs, 4 * sizeof(double));
 	cudaMalloc((void**)&cuTrainingSetOrder, 4 * sizeof(int));
-	cudaMalloc((void**)&cuOutputWeights, numHiddenNodes * numOutputs *  sizeof(float));
+	cudaMalloc((void**)&cuOutputWeights, numHiddenNodes * numOutputs *  sizeof(double));
 
 	// Initialize All The Weights
 	for(int i = 0; i < numInputs; i++){
@@ -97,15 +97,15 @@ int main(int argc, char** argv){
 
 
 	int numTrainingSets = 4; 
-	cudaMemcpy(cuHiddenLayer, hiddenLayer, numHiddenNodes * sizeof(float), cudaMemcpyHostToDevice);
-	cudaMemcpy(cuOutputLayer, outputLayer, numOutputs * sizeof(float), cudaMemcpyHostToDevice);
-	cudaMemcpy(cuHiddenLayerBias, hiddenLayerBias, numHiddenNodes * sizeof(float), cudaMemcpyHostToDevice);
-	cudaMemcpy(cuOutputLayerBias, outputLayerBias, numOutputs * sizeof(float), cudaMemcpyHostToDevice);
-	cudaMemcpy(cuHiddenWeights, hiddenWeights, numInputs * sizeof(float), cudaMemcpyHostToDevice);
-	cudaMemcpy(cuTrainingInputs, training_inputs, 8 * sizeof(float), cudaMemcpyHostToDevice);
-	cudaMemcpy(cuTrainingOutputs, training_outputs, 4 * sizeof(float), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuHiddenLayer, hiddenLayer, numHiddenNodes * sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuOutputLayer, outputLayer, numOutputs * sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuHiddenLayerBias, hiddenLayerBias, numHiddenNodes * sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuOutputLayerBias, outputLayerBias, numOutputs * sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuHiddenWeights, hiddenWeights, numInputs * sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuTrainingInputs, training_inputs, 8 * sizeof(double), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuTrainingOutputs, training_outputs, 4 * sizeof(double), cudaMemcpyHostToDevice);
 	cudaMemcpy(cuTrainingSetOrder, trainingSetOrder, 4 * sizeof(int), cudaMemcpyHostToDevice);
-	cudaMemcpy(cuOutputWeights, outputWeights, numHiddenNodes * sizeof(float), cudaMemcpyHostToDevice);
+	cudaMemcpy(cuOutputWeights, outputWeights, numHiddenNodes * sizeof(double), cudaMemcpyHostToDevice);
 
 
 
@@ -142,20 +142,20 @@ int main(int argc, char** argv){
        	// Specify user inputs	
 
        // Create two pieces of test input
-       float test_input[2] ={atof(argv[1]), atof(argv[2])}; 
+       double test_input[2] ={atof(argv[1]), atof(argv[2])}; 
        
-       float* cuInputs; 
+       double* cuInputs; 
        
-       cudaMalloc((void**)&cuInputs, 2 * sizeof(float));
+       cudaMalloc((void**)&cuInputs, 2 * sizeof(double));
        
-       cudaMemcpy(cuInputs, test_input, 2 * sizeof(float), cudaMemcpyHostToDevice);
+       cudaMemcpy(cuInputs, test_input, 2 * sizeof(double), cudaMemcpyHostToDevice);
        
        forwardFeed<<<1, 1>>>(cuInputs, cuHiddenWeights, cuHiddenLayer, cuOutputLayer, cuOutputWeights, cuOutputLayerBias, cuHiddenLayerBias, numHiddenNodes, numInputs, numOutputs, 0);	
        cudaDeviceSynchronize();
 
        // Transfer the memory off of the GPU to the CPU 
        //outputLayer[0] = 15.00f;
-       cudaMemcpy(outputLayer, cuOutputLayer, numOutputs * sizeof(float), cudaMemcpyDeviceToHost);
+       cudaMemcpy(outputLayer, cuOutputLayer, numOutputs * sizeof(double), cudaMemcpyDeviceToHost);
 
        // Run the forward feed function given the Hidden Weights
